safari-534.53-branch/LayoutTests/storage/indexeddb/transaction-rollback-expected.txt

Test IndexedDB transaction rollback.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitIndexedDB' in window is true
PASS webkitIndexedDB == null is false
webkitIndexedDB.open('transaction-rollback')
openSuccess():
db = event.target.result
request = db.setVersion('version 1')
cleanDatabase():
transaction = event.target.result
Deleted all object stores.
db.createObjectStore('myObjectStore')
PASS db.objectStoreNames.length is 1
setVersionComplete():
transaction = db.transaction(['myObjectStore'], webkitIDBTransaction.READ_WRITE)
store = transaction.objectStore('myObjectStore')
store.add('rollbackValue', 'rollbackKey123')
addSuccess():
PASS event.target.result is 'rollbackKey123'
store.openCursor()
openCursorSuccess():
cursor = event.target.result
abortCallback():
Transaction was aborted.
transaction = db.transaction(['myObjectStore'], webkitIDBTransaction.READ)
store = transaction.objectStore('myObjectStore')
store.get('rollbackKey123')
getSuccess():
PASS event.target.result is undefined
PASS cursor.value is 'rollbackValue'
PASS successfullyParsed is true

TEST COMPLETE
